+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ dev
 - Fix crashing summary and plots for samples with multivariate priors
 - Add progress bar for inference methods
 - Add method save to Sample objects
+- Add support for giving seed to `generate`
+- Implement elfi.plot_params_vs_node for plotting parameters vs. node output
 
 0.7.2 (2018-06-20)
 ------------------